bos2lua: keep large shared values exact instead of rounding through a float
A rules param stores a number as a float, which only holds a whole number
exactly up to 2^24, and a packed map position usually does not fit. Values
outside that range now go into rules params as their decimal string, which
the engine keeps verbatim, and every read converts back with tonumber. The
preview's RuleValue also rounds a plain Lua number through f32, matching the
engine's own float storage, so a test can tell the two paths apart.

+++ crates/coilbox-springlua/src/unitscript.rs
@@ -162,8 +162,10 @@ enum RuleValue {
 impl RuleValue {
     fn of(value: &Value) -> Self {
         match value {
-            Value::Integer(number) => RuleValue::Number(*number as f64),
-            Value::Number(number) => RuleValue::Number(*number),
+            // The engine keeps a rules param number as a float, so this rounds
+            // it through f32 the same way SetRulesParam does.
+            Value::Integer(number) => RuleValue::Number(f64::from(*number as f32)),
+            Value::Number(number) => RuleValue::Number(f64::from(*number as f32)),
             Value::Boolean(flag) => RuleValue::Number(f64::from(u8::from(*flag))),
             Value::String(text) => {
                 let text = text.to_string_lossy();
